settings also now cleaned up
