+++ lib/services/globals.dart
@@ -2,7 +2,7 @@ import 'package:mateapp/models/models.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:mateapp/services/services.dart';
 
-class Global {
+mixin Global {
   static final FirebaseAnalytics analytics = FirebaseAnalytics();
 
   static final Map models = {
