+++ lib/styles/styles.dart
@@ -3,3 +3,4 @@ export 'gradients.dart';
 export 'shadows.dart';
 export 'shapes.dart';
 export 'textstyles.dart';
+export 'themes.dart';
