+++ lib/views/authenticator.dart
@@ -1,0 +1,16 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import 'views.dart';
+
+class Authenticator extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final User user = Provider.of<User>(context);
+    if (user != null) {
+      return Home();
+    }
+    return Welcome();
+  }
+}
