+++ lib/views/settings/local_widgets/local_widgets.dart
@@ -1,0 +1,2 @@
+export 'settings_list.dart';
+export 'settings_list_item.dart';
