+++ lib/views/settings/settings_account/local_widgets/local_widgets.dart
@@ -1,0 +1,5 @@
+export 'settings_account_analytics.dart';
+export 'settings_account_properties.dart';
+export 'settings_account_property.dart';
+export 'settings_account_votediagram.dart';
+export 'settings_account_votediagram_legend.dart';
