+++ lib/views/views.dart
@@ -1,5 +1,6 @@
 export 'cafeteria/cafeteria.dart';
 export 'calendar/calendar.dart';
+export 'home.dart';
 export 'news/news.dart';
 export 'organisation/organisation.dart';
 export 'settings/settings.dart';
