+++ lib/widgets/widgets.dart
@@ -1,5 +1,4 @@
 export 'loading.dart';
-export 'tag.dart';
 export 'sliver_loading_indicator.dart';
 export 'static_navigation_bar.dart';
-export 'settings_list_item.dart';
+export 'tag.dart';
